chore: convert request to TypeScript and to ES6 class
