+++ packages/proxy/test/integration/net-stubbing.spec.ts
@@ -13,7 +13,7 @@ import { DocumentDomainInjection } from '@packages/network-tools'
 import { EventEmitter } from 'events'
 import { RemoteStates } from '@packages/server/lib/remote_states'
 import { CookieJar } from '@packages/server/lib/util/cookies'
-const Request = require('@packages/server/lib/request')
+import { Request as ServerRequest } from '@packages/server/lib/request'
 const getFixture = async () => {}
 
 describe('network stubbing', () => {
@@ -59,7 +59,7 @@ describe('network stubbing', () => {
       getCookieJar: () => new CookieJar(),
       remoteStates,
       getFileServerToken: () => 'fake-token',
-      request: new Request(),
+      request: new ServerRequest(),
       getRenderedHTMLOrigins: () => ({}),
       serverBus: new EventEmitter(),
       getCurrentBrowser: vi.fn(),
